Add null pointer checks
Fixes a crash where, after destroying a nation, active franchies of that
nation would still try to send production on the next turn.

The check was also missing from one of the analogous conversion
functions. I couldn't reproduce the crash there, but thought it best
to add the test anyway.